2020-06-01 Roman 0.9.1.2 - FIX:SYS: The User API function TArrayObj::indexOf() fixed in -1 converting to unsigned. - !!:FIX:DB: TBDS::dataSet() appended for changing-saving the defined entries in the config file         instead complete saving all their in the default DB.     !!:FIX: DB loading is prevented from double loading from the config file and from DB for the equal entries.     !!:FIX: The wrong started DB objects in the first fixed for try them next start tries.     !:FIX: TBDS::{open,close}() fixed for true functional locking by the lock "resTbls". - FIX:DAQ.JavaLikeCalc: Moving the calculation context modification to the option "modifCalc".     FIX: The control interface command "/cntr/cfg/toFunc" some specified. - !!:FIX:UI.Vision: The user manual changing prevented from checking for the lesser or even permition.
git-svn-id: svn://oscada.org/tags/openscada_0.9@2684 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tvariant.cpp
@@ -724,7 +724,7 @@ TVariant TArrayObj::funcCall( const string &id, vector<TVariant> &prms )
 	string sVl = prms[0].getS();
 	unsigned iE = sp;
 	while(iE < mEls.size() && mEls[iE].getS() != sVl) iE++;
-	int rez = (iE < mEls.size()) ? iE : -1;
+	int rez = (iE < mEls.size()) ? (int)iE : -1;
 	dataM.unlock();
 	return rez;
     }
